tests/wpt/tests/webrtc/RTCRtpReceiver-video-jitterBufferTarget-stats.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="timeout" content="long">
<title>Tests RTCRtpReceiver-jitterBufferTarget verified with stats</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/webrtc/RTCPeerConnection-helper.js"></script>
<script src="/webrtc/RTCRtpReceiver-jitterBufferTarget-stats-helper.js"></script>
<body>
<script>
'use strict'

promise_test(async t => {
  await applyJitterBufferTarget(t, "video", 1000);
}, `measure raising and lowering video jitterBufferTarget`);

</script>
</body>
