LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpReceiver-video-jitterBufferTarget-stats-expected.txt


FAIL measure raising and lowering video jitterBufferTarget assert_equals: jitterBufferTarget supported for video expected (object) null but got (undefined) undefined
